src/test/java/ru/job4j/function/FunctionCalculatorTest.java
package ru.job4j.function;

import org.junit.jupiter.api.Test;

import java.util.Arrays;
import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;

class FunctionCalculatorTest {

    @Test
    public void whenLinearFunctionThenLinearResults() {
        FunctionCalculator function = new FunctionCalculator();
        List<Double> result = function.diapason(5, 8, x -> 2 * x + 1);
        List<Double> expected = Arrays.asList(11D, 13D, 15D);
        assertThat(result).containsAll(expected);
    }

    @Test public void  whenQuadraticFunctionThenQuadraticResults() {
        FunctionCalculator function = new FunctionCalculator();
        List<Double> result = function.diapason(2, 6, x -> 3 * Math.pow(x, 2) - 1 * x);
        List<Double> expected = Arrays.asList(10D, 24D, 44D, 70D);
        assertThat(result).containsAll(expected);
    }

    @Test public void  whenExponentialFunctionThenExponentialResults() {
        FunctionCalculator function = new FunctionCalculator();
        List<Double> result = function.diapason(2, 5, x -> Math.pow(4, x));
        List<Double> expected = Arrays.asList(16D, 64D, 256D);
        assertThat(result).containsAll(expected);
    }

}